# WPD-Auth/swagger.yaml
swagger: '2.0'
info:
  description: 'This is a sample JWT authentication service. You can find out more about JWT at [https://jwt.io/](https://jwt.io/). For this sample, you can use the `admin` or `client` users (password: admin and client respectively) to test the authorization filters. Once you have successfully logged in and obtained the token, you should click on the right top button `Authorize` and introduce it with the prefix "Bearer ".'
  version: 1.0.4
  title: The authenticator for the Waterproofing Data (WPD) Work Packages
  contact:
    email: igsd@warwick.ac.uk
  license:
    name: Apache-2.0 License
    url: https://opensource.org/licenses/Apache-2.0
host: localhost:8080
basePath: /
...
paths:
  /forgotpasswords/findallforgotpasswordquestions:
    get:
      tags:
        - forgotpasswords
      summary: ${ForgotPasswordsController.findallforgotpasswordquestions}
      operationId: findallforgotpasswordquestionsUsingGET
      produces:
        - '*/*'
      responses:
        '200':
          description: OK
          schema:
            type: array
            items:
              $ref: '#/definitions/ForgotPasswordsQuestionsResponseDTO'
        '400':
          description: Something went wrong
      security:
        - Authorization:
            - global
      deprecated: false
  /forgotpasswords/loginbyusernameandanswers:
    post:
      tags:
        - forgotpasswords
      summary: ${ForgotPasswordsController.loginbyusernameandanswers}
      operationId: loginbyusernameandanswersUsingPOST
      consumes:
        - application/json
      produces:
        - '*/*'
      parameters:
        - in: body
          name: answers
          description: Answers
          required: false
          schema:
            type: array
            items:
              $ref: '#/definitions/ForgotPasswordsQuestionsUsersAnswersRequestDTO'
        - name: username
          in: query
          description: Username
          required: false
          type: string
          allowEmptyValue: false
      responses:
        '200':
          description: OK
          schema:
            type: string
        '400':
          description: Something went wrong
        '404':
          description: The user or answer don't exist
        '422':
          description: Invalid answers supplied to login. Must have at least 2 correct ones.
      security:
        - Authorization:
            - global
      deprecated: false
  /forgotpasswords/loginbyusernameandkey:
    post:
      tags:
        - forgotpasswords
      summary: ${ForgotPasswordsController.loginbyusernameandkey}
      operationId: loginbyusernameandkeyUsingPOST
      consumes:
        - application/json
      produces:
        - '*/*'
      parameters:
        - name: key
          in: query
          description: Key
          required: false
          type: string
          allowEmptyValue: false
        - name: username
          in: query
          description: Username
          required: false
          type: string
          allowEmptyValue: false
      responses:
        '200':
          description: OK
          schema:
            type: string
        '400':
          description: Something went wrong
        '422':
          description: Invalid username/password supplied
      security:
        - Authorization:
            - global
      deprecated: false
  /forgotpasswords/passwordupdatebyusername:
    post:
      tags:
        - forgotpasswords
      summary: ${ForgotPasswordsController.passwordupdatebyusername}
      operationId: passwordupdatebyusernameUsingPOST
      consumes:
        - application/json
      produces:
        - '*/*'
      parameters:
        - name: newPassword
          in: query
          description: NewPassword
          required: false
          type: string
          allowEmptyValue: false
        - name: username
          in: query
          description: Username
          required: false
          type: string
          allowEmptyValue: false
      responses:
        '200':
          description: OK
        '400':
          description: Something went wrong
        '403':
          description: Access denied
        '404':
          description: The user doesn't exist
        '500':
          description: Expired or invalid JWT token
      security:
        - Authorization:
            - global
        - apiKey: []
      deprecated: false
  /forgotpasswords/saveforgotpasswordquestionsusersanswers:
    post:
      tags:
        - forgotpasswords
      summary: ${ForgotPasswordsController.saveforgotpasswordquestionsusersanswers}
      operationId: saveforgotpasswordquestionsusersanswersUsingPOST
      consumes:
        - application/json
      produces:
        - '*/*'
      parameters:
        - in: body
          name: answer
          description: Forgot Password Questions Users Answers
          required: false
          schema:
            $ref: '#/definitions/ForgotPasswordsQuestionsUsersAnswersRequestDTO'
      responses:
        '200':
          description: OK
        '400':
          description: Something went wrong
        '403':
          description: Access denied
        '404':
          description: The user or forgot password questions don't exist
        '500':
          description: Expired or invalid JWT token
      security:
        - Authorization:
            - global
        - apiKey: []
      deprecated: false
  /forgotpasswords/sendkeybyemail:
    post:
      tags:
        - forgotpasswords
      summary: ${ForgotPasswordsController.sendkeybyemail}
      operationId: sendkeybyemailUsingPOST
      consumes:
        - application/json
      produces:
        - '*/*'
      parameters:
        - name: email
          in: query
          description: Email
          required: false
          type: string
          allowEmptyValue: false
        - name: username
          in: query
          description: Username
          required: false
          type: string
          allowEmptyValue: false
      responses:
        '200':
          description: OK
        '400':
          description: Something went wrong
        '422':
          description: Invalid email supplied
      security:
        - Authorization:
            - global
      deprecated: false
...
securityDefinitions:
  Authorization:
    type: apiKey
    name: Authorization
    in: header
definitions:
...
  ForgotPasswordsQuestionsResponseDTO:
    type: object
    properties:
      id:
        type: integer
        format: int32
      question:
        type: string
      active:
        type: integer
        format: int32
    title: ForgotPasswordsQuestionsResponseDTO
  ForgotPasswordsQuestionsUsersAnswersRequestDTO:
    type: object
    properties:
      id:
        type: integer
        format: int32
      forgotpasswordquestionsid:
        type: integer
        format: int32
      usersid:
        type: integer
        format: int32
      answer:
        type: string
    title: ForgotPasswordsQuestionsUsersAnswersRequestDTO
...